refactor : 전체적인 코드 수정
