+++ 이승창/RestApi/src/main/java/com/example/restapi/app/dto/CartRequestDto.java
@@ -1,9 +1,14 @@
 package com.example.restapi.app.dto;
 
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
 @Getter
 public class CartRequestDto {
+    @NotBlank(message = "상품명은 필수입니다")
     private String item;
+
+    @NotNull(message = "수량은 필수입니다.")
     private int quantity;
 }
