+++ 이승창/RestApi/src/main/java/com/example/restapi/global/code/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.example.restapi.global.exception.code;
+package com.example.restapi.global.code;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
